requirements.txt
pillow
cryptography
numpy
scipy